+++ R/croc-package.R
@@ -1,0 +1,7 @@
+#' @keywords internal
+"_PACKAGE"
+
+
+#' @useDynLib croc, .registration = TRUE
+#' @importFrom Rcpp sourceCpp
+NULL